// mvp/types.ts
export type ExpenseCategory =
  | 'Salaries'
  | 'Maintenance'
  | 'Operations'
  | 'Marketing'
  | 'Rent'
  | 'Utilities'
  | 'Other';
export type ExpenseStatus = 'Paid' | 'Unpaid';
export type ExpensePaymentMethod = 'Cash' | 'Bank Transfer' | 'Card';

export interface Expense {
  id: string; // Expense Number
  category: ExpenseCategory;
  responsible: string; // Employee or Department
  amount: number;
  date: string; // ISO Date String
  status: ExpenseStatus;
  paymentMethod: ExpensePaymentMethod;
  notes?: string;
}